+++ app/Http/Controllers/InvoicesController.php
@@ -10,7 +10,7 @@ class InvoicesController extends Controller
 {
     public function index()
     {
-        $invoices = Invoice::with('customer')->get();
+        $invoices = Invoice::with('customer')->where('user_id', auth()->id())->get();
         return view('invoices.index', ['invoices' => $invoices]);
     }
 
@@ -21,8 +21,13 @@ class InvoicesController extends Controller
 
     public function edit($id)
     {
-        $invoice = Invoice::find($id);
-        return view('invoices.edit', ['invoice' => $invoice]);
+        $invoice = Invoice::where('id', $id)->where('user_id', auth()->id())->first();
+        if($invoice) {
+            return view('invoices.edit', ['invoice' => $invoice]);
+        } else {
+            abort(403);
+            return;
+        }
     }
 
     public function store(InvoiceStoreRequest $request)
@@ -33,6 +38,7 @@ class InvoicesController extends Controller
         $invoice->date = $request->date;
         $invoice->total = $request->total;
         $invoice->customer_id= $request->customer;
+        $invoice->user_id = auth()->id();
 
         $invoice->save();
 
@@ -42,7 +48,12 @@ class InvoicesController extends Controller
 
     public function update(Request $request, $id)
     {
-        $invoice = Invoice::find($id);
+        $invoice = Invoice::where('id', $id)->where('user_id', auth()->id())->first();
+
+        if(!$invoice) {
+            abort(403);
+            return;
+        }
 
         $invoice->number = $request->number;
         $invoice->date = $request->date;
@@ -55,9 +66,12 @@ class InvoicesController extends Controller
 
     public function delete(Request $request, $id)
     {
-        Invoice::destroy($id);
-        return redirect()->route('invoices.index')->with('message', 'Faktura zostałą usunięta');
-
+        if(Invoice::where('id', $id)->where('user_id', auth()->id())->first()) {
+            Invoice::destroy($id);
+            return redirect()->route('invoices.index')->with('message', 'Faktura zostałą usunięta');
+        } else {
+            abort(403);
+            return;
+        }
     }
-
 }
